Rename package refactor to send
Missed some occurences.
Rename `life.qbic.business.notification.refactor` to `life.qbic.business.notification.send`
