+++ sample-notificator-app/src/test/groovy/life/qbic/samplenotificator/components/email/html/HtmlEmailGeneratorSpec.groovy
@@ -1,8 +1,8 @@
 package life.qbic.samplenotificator.components.email.html
 
 import life.qbic.business.notification.create.NotificationContent
-import life.qbic.business.notification.refactor.EmailGenerator
-import life.qbic.business.notification.refactor.NotificationEmail
+import life.qbic.business.notification.send.EmailGenerator
+import life.qbic.business.notification.send.NotificationEmail
 import life.qbic.business.notification.unsubscription.UnsubscriptionLinkSupplier
 import spock.lang.Specification
